Unify JavaScript Proxy implementations (#295)
* `getJavaScriptProxy` no longer requires a `propertyName`.

* create JavaScriptProxyableComponentModelDelegate
