--- modules/charts/gateway/src/main/kotlin/com/mussonindustrial/ignition/embr/charts/component/chart/ApexChartsModelDelegate.kt
@@ -1,38 +1,0 @@
-package com.mussonindustrial.ignition.embr.charts.component.chart
-
-import com.inductiveautomation.ignition.common.util.LogUtil
-import com.inductiveautomation.perspective.gateway.api.Component
-import com.inductiveautomation.perspective.gateway.api.ComponentModelDelegate
-import com.inductiveautomation.perspective.gateway.api.ScriptCallable
-import com.inductiveautomation.perspective.gateway.messages.EventFiredMsg
-import com.mussonindustrial.embr.perspective.gateway.component.ComponentDelegateJavaScriptProxy
-
-class ApexChartsModelDelegate(component: Component) : ComponentModelDelegate(component) {
-
-    private val logger = LogUtil.getModuleLogger("embr-charts", "ApexChartsModelDelegate")
-    private val proxies = hashMapOf<String, ComponentDelegateJavaScriptProxy>()
-
-    override fun onStartup() {
-        component.mdc { logger.debugf("Startup") }
-    }
-
-    override fun onShutdown() {
-        component.mdc { logger.debugf("Shutdown") }
-    }
-
-    override fun handleEvent(message: EventFiredMsg) {
-        proxies.values.forEach {
-            if (it.handles(message)) {
-                it.handleEvent(message)
-            }
-        }
-    }
-
-    @ScriptCallable
-    @Suppress("unused")
-    fun getJavaScriptProxy(property: String): ComponentDelegateJavaScriptProxy {
-        return proxies.getOrPut(property) {
-            ComponentDelegateJavaScriptProxy(component, this, property)
-        }
-    }
-}
